Woo version 1.0
Now with a solid blue background in the icon (I swear i want to make a better one)
Technically this is the second update of this mod, but i forgot to add it to github until like 8 minutes ago.
Anyway, new version lets you disable the item sprite before the tooltip, that's about it iirc.
Oh yeah, apparently there was something about tooltip flashing that i added, didn't seem to have any effect though, might be for something else.